backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
pymongo
yfinance
requests
python-dotenv
aiohttp
transformers
torch
tensorflow
scikit-learn
pandas
numpy
